Loader: extract interactive prompts into prompts.rs
Pulls confirm(), prompt(), format_first_row(), InFlightMode, and
resolve_in_flight_skips() out of main.rs into a dedicated module. These
are all about asking the user questions on stderr/stdin and don't belong
in the dispatch loop's file. main.rs shrinks by ~70 lines.
